streamlit/requirements.txt
streamlit
pillow
keras
tensorflow